refactor: change the structure of the modules files
