+++ data/test-books.js
@@ -1,0 +1,24 @@
+// Test json data with books information
+const books = [
+    {
+        title: "JavaScript: The Good Parts: The Good Parts",
+        author: 'Douglas Crockford',
+        imageLink: 'https://d1b14unh5d6w7g.cloudfront.net/0596517742.01.S001.LXXXXXXX.jpg?Expires=1668162422&Signature=RREeJmwGcI3STAXkIN0ecAVoE-nj63fkoKRxYZu8PDdCQHs6gUMv8udLZM7M~SOm5SPipZHrMxwln0MqJ3q3lgE~fCpuxi9EfgUt4vjKAndAUHV9G2ytSJqVz5tXKqMsOZ7BSSeoHU~tE0tithbqLU1Z03t2kusxTTMCKPdjL2w_&Key-Pair-Id=APKAIUO27P366FGALUMQ',
+        price: 30
+    },
+    {
+        title: "Effective JavaScript: 68 Specific Ways to Harness the Power of JavaScript",
+        author: 'Bradley Meck Alex Young and Mike Cantelon',
+        imageLink: 'https://d1b14unh5d6w7g.cloudfront.net/0321812182.01.S001.LXXXXXXX.jpg?Expires=1668161926&Signature=e8Z3ZMNkakH~-M~0zzx2iFMPj6X0KtWlsqYXs-uBZEz3MMhgixMuUIfmTbVeia3yW0nCo2FO8ynj43azbR0SUem2uX7L7LVR~kKNIYeEW41Xg5WcHjHEw5WTAXD0TL6bSGHpl32MpqTwhl7huyxi9RvBhdREvB85eSnU4Hmx8MY_&Key-Pair-Id=APKAIUO27P366FGALUMQ',
+        price: 44
+
+    },
+    {
+        title: "JavaScript: The Definitive Guide",
+        author: 'David Flanagan',
+        imageLink: 'https://d1b14unh5d6w7g.cloudfront.net/1491950293.01.S001.LXXXXXXX.jpg?Expires=1668161997&Signature=USbHO4ZAES7-r1ytNbt2vx8P4PmT-W4al2U9zs6mBcOaa87qTJOuF9nmHlefwqDvH7CrafM0jw2YKogRkAtqypYhPPi8uz03-v70uzC2DVmIRw~2QWGLi24J5Q243TuJBb4Y5u5S8IUTlUK5PardV8fyCb9Uu-PEYkUzC53ofrE_&Key-Pair-Id=APKAIUO27P366FGALUMQ',
+        price: 120
+    }
+];
+
+export { books };
