+++ index.html
@@ -12,6 +12,6 @@
 </head>
 <body>
     <div id="page"></div>
-    <script src="script.js"></script>
+    <script type="module" src="script.js"></script>
 </body>
 </html>
