+++ modules/footer.js
@@ -1,0 +1,8 @@
+// Create Footer elements:
+const createFooterElements = () => {
+    const tagFooter = document.createElement('footer');
+
+    return tagFooter;
+}
+
+export { createFooterElements };
