+++ styles/catalog-style.css
@@ -1,0 +1,16 @@
+/*  CATALOG SECTION STYLE */
+main {
+    width: 100%;
+    height: 100%;
+}
+
+.catalog-container {
+    max-width: 100%;
+    margin: 18px 0 30px;
+}
+
+.catalog-content {
+    display: flex;
+    justify-content: center;
+    gap: 40px;
+}
