+++ styles/header-style.css
@@ -1,0 +1,28 @@
+/* HEADER STYLE */
+.header-container {
+    width: 1200px;
+    height: auto;
+    padding: 0 15px;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
+
+h1 {
+    font-family: 'Inter';
+    font-style: normal;
+    font-weight: 700;
+    font-size: 24px;
+    line-height: 29px;
+    text-transform: uppercase;
+
+    margin: 24px 0;
+    color: #000000;
+}
+
+.header-divider {
+    width: 100%;
+    height: 2px;
+    background: #393939;
+}
